requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
gunicorn==21.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
